requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
colorama==0.4.6
coverage==7.13.4
daphne==4.2.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.8.0
frozenlist==1.8.0
geoip2==5.2.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
maxminddb==3.1.1
msgpack==1.1.2
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pyotp==2.9.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==8.2
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.23.0
